Get VmId from registry
This is based on the following comments on GitHub:

- https://github.com/mintty/wsltty/issues/356#issuecomment-2394908686
  (this one's mostly to figure out some basic C++ stuff)
- https://github.com/Biswa96/wslbridge2/issues/42#issuecomment-2395382080
  (this one contains the actual fix)

. Take this with a grain of salt, because I know absolutely nothing
about the inner workings of WSL or why the VmId is important at all. All
I want to do is make wsltty work again on my machine.

I'm also everything but a C++ expert, which might explain why I added
the first link above.